requirements-dev.txt
-r requirements.txt
ipython==5.1.0
django-extensions==1.7.4
notebook==4.2.2
